BL-5772 Area outside picture in full screen picture should be black
This is one part of a change. A stylesheet change in Bloom makes the whole image container black when animating. Here we need to force the animation wrapper to have a white background.
I also kept a change which didn't turn out to help which should allow the correct size of image to be determined when it is represented as a child image element instead of a background image.